% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimipo.R
\docType{class}
\name{mimipo}
\alias{mimipo}
\title{Multiply Imputed Pooled Outcome}
\description{
\code{mimipo} object contains data of multiply imputed pooled outcome. \code{mimipo} objects are generated by calls to \code{pool()}.
}
\details{
\code{mimipo} objects has methods for the \code{print()} and \code{summary()} functions (please see \pkg{mice} package reference manual for details).
}
\note{
The \pkg{MatchThem} package does not use the S4 class definitions and instead relies on the S3 list equivalents.
}
\references{
Stef van Buuren and Karin Groothuis-Oudshoorn (2011). \code{mice}: Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of Statistical Software}, 45(3): 1-67. \doi{10.18637/jss.v045.i03}
}
\seealso{
\code{\link[=pool]{pool}}
}
\author{
Farhad Pishgar
}
\keyword{classes}
