% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimira.R
\docType{class}
\name{mimira}
\alias{mimira}
\title{Multiply Imputed Repeated Analyses}
\description{
\code{mimira} object contains data of multiply imputed repeated analyses. \code{mimira} objects are generated by calls to \code{with()}.
}
\details{
\code{mimira} objects has methods for the \code{print()} and \code{summary()} functions (please see \pkg{mice} package reference manual for details).
}
\note{
The \pkg{MatchThem} package does not use the S4 class definitions and instead relies on the S3 list equivalents.
}
\references{
Stef van Buuren and Karin Groothuis-Oudshoorn (2011). \code{mice}: Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of Statistical Software}, 45(3): 1-67. \doi{10.18637/jss.v045.i03}
}
\seealso{
\code{\link[=with]{with()}}
}
\author{
Farhad Pishgar
}
\keyword{classes}
