% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osteoarthritis.R
\docType{data}
\name{osteoarthritis}
\alias{osteoarthritis}
\title{Data of 2,585 Participants in the OAI Project}
\format{This dataset contains 2,585 rows and 7 columns. Each row presents data of an observation (individual) and each column presents data of a characteristics of that observation. The columns are:
\describe{
  \item{AGE}{Age of each observation (individual);}
  \item{SEX}{Gender of each observation (individual), coded as \code{0} (female) and \code{1} (male);}
  \item{BMI}{Estimated body mass index of each observation (individual);}
  \item{RAC}{Race of each observation (individual), coded as \code{0} (Other), \code{1} (Caucasian), \code{2} (African American), and \code{3} (Asian);}
  \item{SMK}{The smoking status of each observation, coded as \code{0} (non-smoker) and \code{1} (smoker)}
  \item{OSP}{Osteoporosis status of each observation (individual) at baseline, coded as \code{0} (negative) and \code{1} (positive); and}
  \item{KOA}{Knee osteoarthritis status of each observation (individual) in the follow-up, coded as \code{0} (at risk) and \code{1} (diagnosed).}
}}
\source{
The information presented in the \code{osteoarthritis} dataset is based on the publicly available data of the Osteoarthritis Initiative (OAI) project (see \url{https://nda.nih.gov/oai/} for details), with changes.
}
\usage{
osteoarthritis
}
\description{
The dataset \code{osteoarthritis} includes demographic data of 2,585 observations (individuals) with or at risk of knee osteoarthritis. The recorded data has missing values in body mass index (\code{BMI}, a quantitative variable), race (\code{RAC}, a categorical qualitative variable), smoking status (\code{SMK}, a binary qualitative variable), osteoporosis status (\code{OSP}, a binary qualitative variable), and knee osteoarthritis status (\code{KOA}, a binary qualitative variable).
}
\keyword{datasets}
