% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wimids.R
\docType{class}
\name{wimids}
\alias{wimids}
\title{Weighted Multiply Imputed Dataset}
\description{
The \code{wimids} object contains data of weighted multiply imputed datasets. The \code{wimids} object is generated by the \code{weightthem()} function.
}
\details{
The \code{wimids} objects has methods for the \code{print()} and \code{summary()} functions.
}
\note{
The \pkg{MatchThem} package does not use the S4 class definitions and instead relies on the S3 list equivalents.
}
\references{
Stef van Buuren and Karin Groothuis-Oudshoorn (2011). \code{mice}: Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of Statistical Software}, 45(3): 1-67. \url{https://www.jstatsoft.org/v45/i03/}
}
\seealso{
\code{\link[=weightthem]{weightthem}}
}
\author{
Farhad Pishgar
}
