% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wimids.R
\docType{class}
\name{wimids}
\alias{wimids}
\title{Weighted Multiply Imputed Datasets}
\description{
\code{wimids} object contains data of weighted multiply imputed datasets. The \code{wimids} object is generated by calls to the \code{weightthem()}.
}
\details{
\code{wimids} objects have methods for \code{print()}, \code{summary()}, and \code{\link[=cbind.mimids]{cbind()}}.
}
\note{
The \pkg{MatchThem} package does not use the S4 class definitions and instead relies on the S3 list equivalents.
}
\references{
Stef van Buuren and Karin Groothuis-Oudshoorn (2011). \code{mice}: Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of Statistical Software}, 45(3): 1-67. \doi{10.18637/jss.v045.i03}
}
\seealso{
\code{\link[=weightthem]{weightthem()}}, \code{\link[=as.wimids]{as.wimids()}}
}
\author{
Farhad Pishgar
}
\keyword{classes}
